print first 98 fibo nums. last test

+++ functions_nested_loops/104-fibonacci.c
@@ -7,7 +7,7 @@
 int main(void)
 {
 	int max = 96, term;
-	unsigned long fib, fib_term1 = 1, fib_term2 = 2;
+	unsigned long long fib, fib_term1 = 1, fib_term2 = 2;
 
 	printf("%lu, %lu, ", fib_term1, fib_term2);
 	for (term = 0; term < max; term++)
